django-rest-base/requirements.txt
<<<<<<< HEAD
asgiref==3.8.1
Django==4.2.13
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==24.2
django-jazzmin==3.0.0
django-model-utils==4.5.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-excel==2.4.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
openpyxl==3.1.2
packaging==24.0
psycopg==3.1.18
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
=======
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.35.65
botocore==1.35.65
build==1.2.2.post1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.3
dill==0.3.6
distlib==0.3.9
Django==5.1.3
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.1
django-model-utils==5.0.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnslib==0.9.25
dnspython==2.7.0
drf-excel==2.5.1
drf-yasg==1.21.8
ecdsa==0.19.0
et_xmlfile==2.0.0
filelock==3.16.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.2
localstack==3.8.1
localstack-core==3.8.1
localstack-ext==3.8.1
markdown-it-py==3.0.0
mdurl==0.1.2
myenv==201902.7
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
panda==0.3.1
pandas==2.2.3
platformdirs==4.3.6
plux==1.12.1
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.0
pyotp==2.9.0
pyproject_hooks==1.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.10.3
semver==3.0.2
six==1.16.0
sqlparse==0.5.2
tabulate==0.9.0
tailer==0.4.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.27.1
wcwidth==0.2.13
>>>>>>> 6f4e717 (Update submodule)
